Move FacilitySearchUsername serializer and viewset to auth/viewsets/facility_user.py
Pure move; public/api_urls.py imports it from the new module.
