+++ kolibri/core/auth/viewsets/facility_user.py
@@ -28,6 +28,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
 from kolibri.core import error_constants
+from kolibri.core.api import BaseValuesViewset
 from kolibri.core.api import ReadOnlyValuesViewset
 from kolibri.core.api import ValuesViewset
 from kolibri.core.api import ValuesViewsetOrderingFilter
@@ -365,6 +366,12 @@ class PublicFacilityUserSerializer(serializers.ModelSerializer):
         )
 
 
+class FacilitySearchUsernameSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = FacilityUser
+        fields = ("id", "username")
+
+
 class PublicFacilityUserViewSet(ReadOnlyValuesViewset):
     queryset = FacilityUser.objects.all().order_by("id")
     serializer_class = PublicFacilityUserSerializer
@@ -392,6 +399,54 @@ class PublicFacilityUserViewSet(ReadOnlyValuesViewset):
             queryset = queryset.filter(id=self.request.user.id)
 
         return queryset
+
+
+class FacilitySearchUsernameViewSet(BaseValuesViewset):
+    filter_backends = (DjangoFilterBackend, filters.SearchFilter)
+    filterset_fields = ("facility",)
+    search_fields = ("^username",)
+
+    serializer_class = FacilitySearchUsernameSerializer
+
+    def list(self, request, *args, **kwargs):
+        facility_id = request.query_params.get("facility", None)
+        if facility_id is None:
+            content = "Missing parameter: facility is required"
+            return Response(content, status=status.HTTP_412_PRECONDITION_FAILED)
+        try:
+            facility = Facility.objects.get(id=facility_id)
+        # A malformed facility UUID raises ValueError, not DoesNotExist (#11735).
+        except (AttributeError, Facility.DoesNotExist, ValueError):
+            content = "The facility does not exist in this device"
+            return Response(content, status=status.HTTP_404_NOT_FOUND)
+
+        if facility.dataset.learner_can_login_with_no_password:
+            queryset = self.filter_queryset(self.get_queryset())
+            return Response(self.serialize(queryset))
+        else:
+            # A facility that requires a password must not disclose user ids, nor
+            # let a prefix search enumerate its usernames (#9512).
+            username = request.query_params.get("search", None)
+            queryset = self.get_queryset().filter(
+                facility=facility_id,
+                # Usernames are unique case-insensitively (#11739).
+                username__iexact=username,
+            )
+            response = (
+                [
+                    {"username": username, "id": None},
+                ]
+                if queryset
+                else []
+            )
+            return Response(response)
+
+    def get_queryset(self):
+        # Learners only; the isnull arm catches users with no DevicePermissions
+        # row at all, which the is_superuser=False arm never matches (#9512).
+        return FacilityUser.objects.filter(roles=None).filter(
+            Q(devicepermissions__is_superuser=False) | Q(devicepermissions__isnull=True)
+        )
 
 
 class FacilityUserViewSet(ValuesViewset, BulkDeleteMixin):
